+++ src/MyLab.Search.EsTest/EsFixture.cs
@@ -1,10 +1,11 @@
 ﻿using System.Threading.Tasks;
 using Elasticsearch.Net;
+using MyLab.Search.EsAdapter;
 using Nest;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace MyLab.Elastic.Test
+namespace MyLab.Search.EsTest
 {
     /// <summary>
     /// Provides manager to work with remote ES instance
@@ -13,9 +14,10 @@ namespace MyLab.Elastic.Test
         where TConnectionProvider : IConnectionProvider, new()
     {
         private readonly IConnectionPool _connection;
-        private readonly ElasticClient _client;
 
         public IEsManager Manager { get; private set; }
+
+        public ElasticClient EsClient { get; }
 
         /// <summary>
         /// Test output. Set to get logs.
@@ -45,12 +47,22 @@ namespace MyLab.Elastic.Test
                 Output?.WriteLine(ApiCallDumper.ApiCallToDump(details));
             });
 
-            _client = new ElasticClient(settings);
+            EsClient = new ElasticClient(settings);
+        }
+
+        public IEsSearcher<TDoc> CreateSearcher<TDoc>() where TDoc : class
+        {
+            return new EsSearcher<TDoc>(new SingleEsClientProvider(EsClient), null, (ElasticsearchOptions) null);
+        }
+
+        public IEsIndexer<TDoc> CreateIndexer<TDoc>() where TDoc : class
+        {
+            return new EsIndexer<TDoc>(new SingleEsClientProvider(EsClient), null, (ElasticsearchOptions)null);
         }
 
         public Task InitializeAsync()
         {
-            Manager = new EsManager(new SingleEsClientProvider(_client), (ElasticsearchOptions)null);
+            Manager = new EsManager(new SingleEsClientProvider(EsClient), (ElasticsearchOptions)null);
 
             return Task.CompletedTask;
         }