+++ src/MyLab.Search.EsTest/EsIndexFactoryFixture.cs
@@ -1,10 +1,11 @@
 ﻿using System;
 using System.Threading.Tasks;
 using Elasticsearch.Net;
+using MyLab.Search.EsAdapter;
 using Nest;
 using Xunit.Abstractions;
 
-namespace MyLab.Elastic.Test
+namespace MyLab.Search.EsTest
 {
     /// <summary>
     /// CreateAsync tmp index factory with specified model mapping
@@ -62,9 +63,9 @@ namespace MyLab.Elastic.Test
         {
             await using var indexLife = await TmpIndexLife<TDoc>.CreateAsync(_client, tmpIndexName);
 
-            var searcher = new EsSearcher<TDoc>(new SingleEsClientProvider(_client), null).ForIndex(indexLife.IndexName);
+            var searcher = new EsSearcher<TDoc>(new SingleEsClientProvider(_client), null, (ElasticsearchOptions)null).ForIndex(indexLife.IndexName);
             var manager = new EsManager(new SingleEsClientProvider(_client), (ElasticsearchOptions)null);
-            var indexer = new EsIndexer<TDoc>(new SingleEsClientProvider(_client), null).ForIndex(indexLife.IndexName);
+            var indexer = new EsIndexer<TDoc>(new SingleEsClientProvider(_client), null, (ElasticsearchOptions)null).ForIndex(indexLife.IndexName);
 
             await action(new TestServices<TDoc>
             {
@@ -82,9 +83,9 @@ namespace MyLab.Elastic.Test
         {
             await using var indexLife = await TmpIndexLife<TDoc>.CreateAsync(_client, tmpIndexName);
 
-            var searcher = new EsSearcher<TDoc>(new SingleEsClientProvider(_client), null).ForIndex(indexLife.IndexName);
+            var searcher = new EsSearcher<TDoc>(new SingleEsClientProvider(_client), null, (ElasticsearchOptions)null).ForIndex(indexLife.IndexName);
             var manager = new EsManager(new SingleEsClientProvider(_client), (ElasticsearchOptions)null);
-            var indexer = new EsIndexer<TDoc>(new SingleEsClientProvider(_client), null).ForIndex(indexLife.IndexName);
+            var indexer = new EsIndexer<TDoc>(new SingleEsClientProvider(_client), null, (ElasticsearchOptions)null).ForIndex(indexLife.IndexName);
 
             return await func(new TestServices<TDoc>
             {